DO NOT MERGE ANYWHERE: Fix P2P failure when setting WPS P2P IE
BRCM kernel has incorrect handling of the command to set WPS IE in
probe response. Regardless of the size of the vendor specific IE,
the kernel hangs.

P2P functionality doesn't need WPS IE to be sent in probe response
This CL removes sending of this command the driver to the kernel.

Change-Id: I8bbd12c40503205049129336a8d271b3d52d73da
Bug: 34619601
Test: Manual P2p tests between two devices with BRCM drivers