WiFi: Vendor HAL function to wait for driver ready
This commit adds a vendor hal function to wait for the driver to be
ready. This ensures the driver is ready for operation before framework
starts to use it.

Note: Broadcom mentioned that checking for the presence of
/sys/class/net/wlan0 may not guarantee that WiFi is fully
initialized, and we should check /sys/class/net/wlan0/flags
for IFF_UP instead.

However, this doesn't work. The /sys/class/net/wlan0/flags value
never moves beyond 0x1002 when WiFi is turned on, until after this
function (and its subsequent functions) is invoked.

Bug: 74120746
Change-Id: If8c624c4580d456abcb1437493783b3f114d5961