bcm4343: Update wifi firmware for CE cert
Wi-fi adaptivity test was failed because of CAL signal. After Interference
Signal is stopped. The transmission signal is restricted.  But, some short
signal has occurred.  So, the test equipment considers the short signal as
the short control signal. Finally, Wi-fi adaptivity test was failed.

Even if the transmission signal is restricted for adaptivity procedure of
chipset,  some short signal is found in the testbed.  The reason for this
signal is due to Tx power calibration of internal FW procedure.
This procedure is executed by internal ED parameter  and  detected signal
level in Air condition.  But, in some exceptional cases, this calibration
should be skipped.   The reason for the short signal is that the level of
calibration skip function is not matched with the device condition.

This firmware was fixed to avoid those trigger PHY calibrations when ED is
ON.   The condition of calibration skip function is modified to match with
the device HW condition.

Version: 7.13.53.13

Bug: 78106685
Change-Id: I70150deb311350f93ac275e2210ecd0b836c49b1
(cherry picked from commit ac159a6bb09f9e99827412c138ea874278b5aa4e)
(cherry picked from commit 311f4e08eb37244e6190382cd18db16173182184)