HAL: Fixed vendor HAL crash due to RSSI monitor operation
[Issue]
Vendor HAL crash happened due to use-after-free.
So, segmentation fault happened due to wrong memory access.

[Analysis]
Crash is closely related to RSSI_MONITOR operation. There are 2 threads, the last states are as follows.
- The one thread is registering vendor handler.
- The another one thread is receiving rssi monitor event from host driver.

The situation can be summarized as follows.
-----------------------
1. stopRssiMonitoring failed. So, unregisterVendorHandler(GOOGLE_OUI, GOOGLE_RSSI_MONITOR_EVENT) didn't run.
=> So, 'info->event_cb[]' arrays still has invalidated/older record
2. startRssiMonitoring called. The subsequent registerVendorHandler(GOOGLE_OUI, GOOGLE_RSSI_MONITOR_EVENT) is on-going.
3. At this time, RSSI monitor command was already sent to the driver. And then, received an event message from the driver even though the event handler wasn't registered yet.
=> Due to above #1, invalidated/older record retrieved for corresponding event message. Finally, SIGFAULT detected by invalid address accessing.
-----------------------

[Solution]
It is hard to 100% figure out the root cause with currently limited information.
Obviously, the vendor handler was not unregistered properly under the specific path and accessed an invalidated address.
Although it is not confirmed in the log, if stopRssiMonitoring() fails after startRssiMonitoring() is normally called,
unregisterVendorHandler(GOOGLE_OUI, GOOGLE_RSSI_MONITOR_EVENT) is not executed.

If then, it can cause this issue. So, we decided to fix like below.
-----------------------
1. Basically, the registerVendorHandler() call would be good to preceded before issuing a command to the driver. (Many classes implemented that way.)
2. wifi_register_handler() fix to do update instead of addition if the same [nl_cmd, vendor_id, vendor_subcmd] record exists.
3. Add the log and adjust log level from VERBOSE to INFO if required.
-----------------------

Bug: 190846707
Test: Verified using halutil, rssi_monitor function works well.
Change-Id: I985ef8a1a5cb8c683b8741da6e498d06f12b07dc